// src/api/user/message/user.message.ts
export enum UserCRUDMessage {
  // Créer
  CREATE_SUCCESS = 'Utilisateur créé avec succès.',
  CREATE_ERROR = "Erreur lors de la création de l'utilisateur.",

  // Lire
  READ_SUCCESS = 'Utilisateur récupéré avec succès.',
  READ_ERROR = "Erreur lors de la récupération de l'utilisateur.",

  // Mettre à jour
  UPDATE_SUCCESS = 'Utilisateur mis à jour avec succès.',
  UPDATE_ERROR = "Erreur lors de la mise à jour de l'utilisateur.",

  // Supprimer
  DELETE_SUCCESS = 'Utilisateur supprimé avec succès.',
  DELETE_ERROR = "Erreur lors de la suppression de l'utilisateur.",
}